abstracting main
